requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.2
Flask-WhooshAlchemy==0.56
Jinja2==2.7.3
MarkupSafe==0.23
Paste==1.7.5.1
SQLAlchemy==0.9.8
Tempita==0.5.2
WTForms==2.0.1
WebOb==1.4
Werkzeug==0.9.6
Whoosh==2.6.0
argparse==1.2.1
blinker==1.3
decorator==3.4.0
itsdangerous==0.24
pbr==0.10.0
pyenchant==1.6.6
pyswip==0.2.3
python-openid==2.2.5
six==1.8.0
sqlalchemy-migrate==0.9.2
sqlparse==0.1.13
webapp2==2.5.2
wsgiref==0.1.2
